feat: delete comment
