+++ apps/mobile/lib/icons/Trash.tsx
@@ -1,4 +1,4 @@
 import { Trash2 } from "lucide-react-native";
 import { iconWithClassName } from "./iconWithClassName";
 iconWithClassName(Trash2);
-export { Trash2 };
+export { Trash2 as Trash };
