+++ apps/mobile/src/app/(tabs)/(feed,home,search,profile)/[handle]/ratings/[id].tsx
@@ -1,34 +1,10 @@
 import NotFoundScreen from "#/app/+not-found";
-import { timeAgo } from "@recordscratch/lib";
 import { FlashList } from "@shopify/flash-list";
-import { Link, Stack, useLocalSearchParams } from "expo-router";
-import { Pressable, View } from "react-native";
+import { Stack, useLocalSearchParams } from "expo-router";
+import { View } from "react-native";
+import { Comment } from "~/components/Comment";
 import { Review } from "~/components/Review";
-import { UserAvatar } from "~/components/UserAvatar";
-import { Text } from "~/components/ui/text";
-import { RouterOutputs, api } from "~/lib/api";
-import { getImageUrl } from "~/lib/image";
-
-export const Comment = ({
-	comment: { content, profile, updatedAt },
-}: {
-	comment: RouterOutputs["comments"]["list"][0];
-}) => {
-	return (
-		<View className="p-4 gap-4">
-			<Link href={`/${String(profile.handle)}`} asChild>
-				<Pressable className="flex flex-row flex-wrap items-center gap-2">
-					<UserAvatar size={40} imageUrl={getImageUrl(profile)} />
-					<Text className="text-lg">{profile.name}</Text>
-					<Text className="text-left text-muted-foreground text-lg">
-						@{profile.handle} • {timeAgo(updatedAt)}
-					</Text>
-				</Pressable>
-			</Link>
-			<Text className="text-lg">{content}</Text>
-		</View>
-	);
-};
+import { api } from "~/lib/api";
 
 const RatingPage = () => {
 	const { id, handle } = useLocalSearchParams<{ id: string; handle: string }>();
